test(calldata): co-locate validation tests
