--- suite-common/calldata/src/tests/validation/evm/validateAddress.test.ts
@@ -1,8 +1,0 @@
-import { validateAddressTestCases } from '../../../fixtures/validation/evm/validateAddress.fixture';
-import { validateAddress } from '../../../validation/evm/address';
-
-describe('validateAddress', () => {
-    it.each(validateAddressTestCases)('$description', ({ input, context, expected }) => {
-        expect(validateAddress(input, 'to', context)).toEqual(expected);
-    });
-});
