--- suite-common/calldata/src/tests/validation/shared/uint16.test.ts
@@ -1,8 +1,0 @@
-import { findUint16OverflowIssueTestCases } from '../../../fixtures/validation/shared/uint16.fixture';
-import { findUint16OverflowIssue } from '../../../validation/shared/uint16';
-
-describe('findUint16OverflowIssue', () => {
-    it.each(findUint16OverflowIssueTestCases)('$description', ({ input, expected }) => {
-        expect(findUint16OverflowIssue(input)).toBe(expected);
-    });
-});
