--- suite-common/calldata/src/tests/validation/tron/address.test.ts
@@ -1,17 +1,0 @@
-import {
-    findSelfAddressIssueTestCases,
-    findTronAddressIssueTestCases,
-} from '../../../fixtures/validation/tron/address.fixture';
-import { findSelfAddressIssue, findTronAddressIssue } from '../../../validation/tron/address';
-
-describe('findTronAddressIssue', () => {
-    it.each(findTronAddressIssueTestCases)('$description', ({ input, expected }) => {
-        expect(findTronAddressIssue(input)).toBe(expected);
-    });
-});
-
-describe('findSelfAddressIssue', () => {
-    it.each(findSelfAddressIssueTestCases)('$description', ({ input, context, expected }) => {
-        expect(findSelfAddressIssue(input, context)).toBe(expected);
-    });
-});
