--- suite-common/calldata/src/tests/validation/tron/validateTronAddress.test.ts
@@ -1,8 +1,0 @@
-import { validateTronAddressTestCases } from '../../../fixtures/validation/tron/validateTronAddress.fixture';
-import { validateTronAddress } from '../../../validation/tron/address';
-
-describe('validateTronAddress', () => {
-    it.each(validateTronAddressTestCases)('$description', ({ input, context, expected }) => {
-        expect(validateTronAddress(input, 'to', context)).toEqual(expected);
-    });
-});
