+++ suite-common/calldata/src/validation/evm/validateAddress.test.ts
@@ -1,0 +1,8 @@
+import { validateAddress } from './address';
+import { validateAddressTestCases } from '../../../mocks/validation/evm/mockValidateAddress';
+
+describe('validateAddress', () => {
+    it.each(validateAddressTestCases)('$description', ({ input, context, expected }) => {
+        expect(validateAddress(input, 'to', context)).toEqual(expected);
+    });
+});
