+++ suite-common/calldata/src/validation/shared/validateBytes32.test.ts
@@ -1,0 +1,8 @@
+import { validateBytes32 } from './bytes32';
+import { validateBytes32TestCases } from '../../../mocks/validation/shared/mockValidateBytes32';
+
+describe('validateBytes32', () => {
+    it.each(validateBytes32TestCases)('$description', ({ input, expected }) => {
+        expect(validateBytes32(input, 'proof')).toEqual(expected);
+    });
+});
